Aplicar o uso do UseContext para utilizar propriedade e funções de outros componentes
