+++ src/components/Header/index.tsx
@@ -17,8 +17,6 @@ export function Header ({ onOpenNewTransactionModal }: HeaderProps){
         <button type="button" onClick={onOpenNewTransactionModal}>
           Nova transação
         </button>
-        
-        
       </Content>      
     </Container>
   )
