+++ src/components/Summary/index.tsx
@@ -1,10 +1,15 @@
-import React from "react";
+import React, {useContext} from "react";
 import incomeImg from '../../assets/income.svg'
 import outcomeImg from '../../assets/outcome.svg'
 import totalImg from '../../assets/total.svg'
+import { TransactionsContext } from "../../TransactionsContext";
 import { Container } from "../Summary/styles";
 
 export function Summary(){
+  const transactions = useContext(TransactionsContext);
+
+  console.log(transactions);
+
   return(
     <Container>
       <div>
