+++ src/components/TransactionsTable/index.tsx
@@ -1,25 +1,11 @@
-import { useEffect, useState } from "react";
-import { api } from "../../services/api";
+import { useContext } from "react";
+import { TransactionsContext } from "../../TransactionsContext";
 import { Container } from "./styles";
 
-interface Transaction {
-  id: number;
-  title: number;
-  amount: number;
-  type: string;
-  category: string;
-  createdAt: string;
-}
+
 
 export function TransactionsTable() {
-  const [transactions, setTransactions] = useState<Transaction[]>([]);
-
-  useEffect(() => {
-    api.get('transactions')
-      .then(response => setTransactions(response.data.transactions))
-  }, []);
-  
-  
+  const transactions = useContext(TransactionsContext);
   
   return(
     <Container>
